requirements_new.txt
# ==============================================================================
# REQUIREMENTS FOR MODERN UNIVERSITY MANAGEMENT SYSTEM
# تاریخ ایجاد: ۱۴۰۳/۰۶/۲۰
# معماری: Django + ابزارهای مکمل
# ==============================================================================

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9  # PostgreSQL
redis==5.0.1  # Redis for caching and sessions
django-redis==5.4.0

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.26.5
drf-yasg==1.21.7

# Image Processing
Pillow==10.1.0

# File Handling
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7

# Data Validation & Serialization
marshmallow==3.20.1
cerberus==1.3.5

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment Management
python-decouple==3.8
python-dotenv==1.0.0

# Task Queue (Background Jobs)
celery==5.3.4
kombu==5.3.4
billiard==4.2.0

# Message Broker
redis==5.0.1  # Can be used with Celery
# rabbitmq-python==1.0.0  # Alternative to Redis

# Monitoring & Logging
sentry-sdk[django]==1.38.0
structlog==23.2.0
django-structlog==6.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development Tools
django-debug-toolbar==4.2.0
django-silk==5.0.4
ipython==8.17.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security
django-cryptography==1.1
cryptography==41.0.7
bcrypt==4.1.2

# Internationalization
django-modeltranslation==0.18.11

# Social Authentication (Optional)
social-auth-app-django==5.4.0

# Admin Interface Enhancement
django-admin-interface==0.28.6
django-grappelli==3.0.8

# API Rate Limiting
django-ratelimit==4.1.0

# Search (if needed later)
django-haystack==3.2.1
elasticsearch==8.11.0

# Time Zone Handling
pytz==2023.3
tzdata==2023.3

# Excel/CSV Export
tablib==3.5.0
django-import-export==3.3.1

# Backup & Migration
django-dbbackup==4.0.2

# Performance Monitoring
django-silk==5.0.4
newrelic==9.2.0  # Optional

# Email
django-anymail==10.2

# File Storage (Cloud)
boto3==1.34.0  # AWS S3
django-storages==1.14.2

# ==============================================================================
# BIG DATA & ANALYTICS STACK (for future implementation)
# ==============================================================================

# Apache Kafka (Python Client)
confluent-kafka==2.3.0
kafka-python==2.0.2

# Apache Parquet
pyarrow==14.0.1
fastparquet==2023.10.1

# Apache Iceberg (Table Format)
pyiceberg==0.5.1

# Stream Processing
pyflink==1.18.0  # Apache Flink Python API

# OLAP Database Clients
clickhouse-driver==0.2.6  # ClickHouse
trino==0.328.0  # Trino/Presto

# Object Storage
minio==7.2.0  # MinIO Python SDK

# Workflow Orchestration
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-redis==3.4.0

# Machine Learning (Optional)
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2

# Data Visualization (Optional)
plotly==5.17.0
bokeh==3.3.2

# ==============================================================================
# DOCKER & DEPLOYMENT
# ==============================================================================

# WSGI/ASGI Servers
gunicorn==21.2.0
uvicorn==0.24.0
daphne==4.0.0

# Container Health Checks
healthcheck==1.3.3

# Configuration Management
django-environ==0.11.2

# ==============================================================================
# DEVELOPMENT & TESTING EXTRAS
# ==============================================================================

# Code Quality
pre-commit==3.5.0
bandit==1.7.5  # Security linting
safety==2.3.5  # Dependency vulnerability check

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Database Fixtures
django-fixtures==0.1.0

# Mock Data Generation
mimesis==12.1.0
hypothesis==6.92.1

# API Testing
httpie==3.2.2
curl-cffi==0.5.10
